Moved components to components folder and created Button component.
